Test with multiple versions of PostgreSQL in Travis

+++ .travis.sh
@@ -2,32 +2,44 @@
 
 set -eu
 
-pgdg_repository() {
+pgdg_repositories() {
 	local sourcelist='sources.list.d/pgdg.list'
 
+	sudo tee "/etc/apt/$sourcelist" <<-repositories
+		deb http://apt.postgresql.org/pub/repos/apt/ $(lsb_release -cs)-pgdg main
+		deb http://apt.postgresql.org/pub/repos/apt/ $(lsb_release -cs)-pgdg-testing main 10
+	repositories
+
 	sudo apt-key adv --keyserver 'hkp://ha.pool.sks-keyservers.net' --recv-keys 'ACCC4CF8'
-	echo deb http://apt.postgresql.org/pub/repos/apt/ trusty-pgdg main | sudo tee "/etc/apt/$sourcelist"
 	sudo apt-get -o Dir::Etc::sourcelist="$sourcelist" -o Dir::Etc::sourceparts='-' -o APT::Get::List-Cleanup='0' update
 }
 
-postgresql_configure() {
-	sudo tee /etc/postgresql/9.1/main/pg_hba.conf > /dev/null <<-config
+postgresql_install() {
+	if [ -z "${PGVERSION:-}" ]; then
+		PGVERSION="$( psql -d postgres -XAtc "select regexp_replace(current_setting('server_version'), '[.][0-9]+$', '')" )"
+	else
+		sudo service postgresql stop
+		xargs sudo apt-get -y --purge remove <<-packages
+			libpq-dev
+			libpq5
+			postgresql
+			postgresql-client-common
+			postgresql-common
+		packages
+		sudo rm -rf /var/lib/postgresql
+	fi
+
+	xargs sudo apt-get -y install <<-packages
+		postgresql-${PGVERSION}
+		postgresql-${PGVERSION}-ip4r
+	packages
+
+	sudo tee /etc/postgresql/${PGVERSION}/main/pg_hba.conf > /dev/null <<-config
 		local  all  all                trust
 		host   all  all  127.0.0.1/32  trust
 	config
 
 	sudo service postgresql restart
-}
-
-postgresql_uninstall() {
-	sudo service postgresql stop
-	xargs sudo apt-get -y --purge remove <<-packages
-		libpq-dev
-		libpq5
-		postgresql
-		postgresql-client-common
-		postgresql-common
-	packages
 }
 
 remote_file() {
